api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.13
botocore==1.38.13
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.16.0
idna==3.6
jmespath==1.0.1
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
qrcode==7.4.2
s3transfer==0.12.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.18
uvicorn==0.25.0
